LevelUpKarte und LevelUpKarte.java in Kartentypen verändert dadurch LvlUpCards.java und LvlUpCardsImpl.java gelöscht
und Variablen in MonsterkarteImpl und RassenKarteImpl private gesetzt
