+++ app/src/main/java/com/example/munchkin/Karte/KartenTypen/LevelUpKarte.java
@@ -1,0 +1,7 @@
+package com.example.munchkin.Karte.KartenTypen;
+
+public interface LevelUpKarte extends Schatzkarte{
+
+    //Übernommen von LvlUpCards
+    int increaceLvlByOnePermanet(int currentLevel);
+}
