--- app/src/main/java/com/example/munchkin/Karte/LvlUpCards.java
@@ -1,6 +1,0 @@
-package com.example.munchkin.Karte;
-
-public interface LvlUpCards extends Karte {
-
-    int increaceLvlByOnePermanet(int currentLevel);
-}
